mid rewrite of parser, changing implementation
